all delete buttons and logic

+++ src/main/java/src/javasqlriskmanager/controllers/userscontrollers/DetalleUsuarioController.java
@@ -10,6 +10,7 @@ import javafx.scene.control.TextField;
 import src.javasqlriskmanager.MainApplication;
 import src.javasqlriskmanager.models.Usuario;
 import src.javasqlriskmanager.singletons.IncidentSingleton;
+import src.javasqlriskmanager.singletons.SesionSingleton;
 import src.javasqlriskmanager.singletons.UserSingleton;
 import src.javasqlriskmanager.utils.ConnectToDB;
 
@@ -78,31 +79,38 @@ public class DetalleUsuarioController implements Initializable {
 
         if(alert.getResult().getText().equals("Aceptar")) {
 
-            try {
-                Connection con = ConnectToDB.connectToDB();
+            if(UserSingleton.getInstance().getUsuario().getID().equals(SesionSingleton.getInstance().getUsuario().getID())) {
+                alert = new Alert(Alert.AlertType.ERROR);
+                alert.setHeaderText(null);
+                alert.setTitle("Error");
+                alert.setContentText("No puedes eliminar el usuario con con el que iniciaste sesión.");
+                alert.showAndWait();
+            } else {
+                try {
+                    Connection con = ConnectToDB.connectToDB();
 
-                String query = "UPDATE Incidents SET ID_CreatorUser = NULL WHERE ID_CreatorUser = ?";
-                PreparedStatement preparedStatement = con.prepareStatement(query);
-                preparedStatement.setLong(1, UserSingleton.getInstance().getUsuario().getID());
-                preparedStatement.execute();
+                    String query = "UPDATE Incidents SET ID_CreatorUser = NULL WHERE ID_CreatorUser = ?";
+                    PreparedStatement preparedStatement = con.prepareStatement(query);
+                    preparedStatement.setLong(1, UserSingleton.getInstance().getUsuario().getID());
+                    preparedStatement.execute();
 
-                query = "UPDATE Incidents SET ID_AssignedUser = NULL WHERE ID_AssignedUser = ?";
-                preparedStatement = con.prepareStatement(query);
-                preparedStatement.setLong(1, UserSingleton.getInstance().getUsuario().getID());
-                preparedStatement.execute();
+                    query = "UPDATE Incidents SET ID_AssignedUser = NULL WHERE ID_AssignedUser = ?";
+                    preparedStatement = con.prepareStatement(query);
+                    preparedStatement.setLong(1, UserSingleton.getInstance().getUsuario().getID());
+                    preparedStatement.execute();
 
-                query = "DELETE FROM Users WHERE ID = ?";
-                preparedStatement = con.prepareStatement(query);
-                preparedStatement.setLong(1, UserSingleton.getInstance().getUsuario().getID());
-                preparedStatement.execute();
+                    query = "DELETE FROM Users WHERE ID = ?";
+                    preparedStatement = con.prepareStatement(query);
+                    preparedStatement.setLong(1, UserSingleton.getInstance().getUsuario().getID());
+                    preparedStatement.execute();
 
-                con.close();
-            } catch (Exception e) {
-                System.out.println(e.getMessage());
-                e.printStackTrace();
+                    con.close();
+                } catch (Exception e) {
+                    System.out.println(e.getMessage());
+                    e.printStackTrace();
+                }
+                irCatalogo();
             }
-
-            irCatalogo();
         }
     }
 }
